using lzma for file compression
